tools: srpmpacker: refactor as a library
